Moving reportError to a separate function
